+++ Runtime/Easing/EasingOptions.cs
@@ -1,0 +1,12 @@
+[System.Serializable]
+public class EasingOptions
+{
+    public EaseType easingType;
+    public int smoothing;
+
+    public EasingOptions(EaseType easingType, int smoothing)
+    {
+        this.easingType = easingType;
+        this.smoothing = smoothing;
+    }
+}